cdsb_platform/pyproject.toml
[project]
name = "cdsb-platform"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "snowflake-connector-python>=3.3.0",
    "streamlit[snowflake]>=1.54.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "curl-cffi>=0.7.0",
]
